% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recode_dict.R
\name{create_recode_dict}
\alias{create_recode_dict}
\title{Create a recoding data dictionary}
\usage{
create_recode_dict(
  df,
  dict_type = "all",
  var_col = "var",
  values_col = "values",
  labels_col = "labels",
  default_col = "default",
  string_col = "string"
)
}
\arguments{
\item{df}{A data.frame or tibble containing unique variable values and value labels
for variables in a dataset. Two optional columns can also be included: `default`, a
default value for old values without a new mapped value/label and `string`, a logical
column indicating whether the new values to be mapped onto the variable are string (1)
or not (0).}

\item{dict_type}{Character string of a type of dictionary to create. Must be one
of `all`, `string`, or `numeric`. Default is `all`.}

\item{var_col}{Character string of the name of the column in `df` that contains unique
variable/column names in your dataset.}

\item{values_col}{A character string specifying the name of the column in `df` that holds
the unique values for each variable/column in your dataset.}

\item{labels_col}{A character string specifying the name of the column in `df` that holds
the unique values/labels to replace old values for each variable/column in your dataset.}

\item{default_col}{A character string specifying the name of the column in `df` that holds
default values to use when a matching (old) value is not found.}

\item{string_col}{A character string specifying the name of the column in `df` that
indicates whether the new values to be mapped onto the variable are string.}
}
\value{
A tibble
}
\description{
Creates a 'recoding' dictionary for string and numeric variables.
Each dictionary includes unique column names from a dataset to recode, a series of
two-sided formulas to define a mapping from old values to new values, default values
to use when a matching (old) value is not found, and a column indicating whether the new
values to be applied to each unique column are string (1) or not (0).
}
