% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_app_data.R
\docType{data}
\name{grad_app}
\alias{grad_app}
\title{Graduate School Application Data}
\format{
A data.frame with 5,000 rows and 14 columns:
\describe{
\item{applicant_id}{(Applicant identification number)}
\item{email}{(email address of an applicant)}
\item{email_preferred}{(preferred email address of applicant)}
\item{school_decision}{(whether a school accepted (A), Waitlisted (W), or
Rejected (R) an applicant)}
\item{student_decision}{(whether a prospective applicant Accepted (A) or
Declined (D) a school's offer)}
\item{ft_pt}{(whether an applicant applied to a Part-time (pt), Full-time (ft),
or would be a Submatriculate (sub))}
\item{gre_verbal}{(applicants' GRE Verbal Reasoning score; values range from 130 - 170)}
\item{gre_quant}{(applicants' GRE Quantitative Reasoning score; values range from 130 -
170)}
\item{gre_analytical}{applicants' GRE Analytical Writing score; values range from
0 to 6, in half-point increments}
\item{applied_dual_prg}{(whether an applicant applied to  dual degree program (YES) or
not (NODUAL))}
\item{birth_date}{(applicants' date of birth (MM/DD/YY))}
\item{gender}{(applicants' gender; one of Woman, Man, or Non-Binary)}
\item{ethnicity}{(applicants' race/ethnicity; one of Hispanic/Latino (HL), Middle Eastern/
Arab (ME), White (W), Black /African American (B), or Asian or Pacific Islander (AP))}
\item{low_income}{(whether an applicant is low income (Y) or not (N))}
}
}
\usage{
grad_app
}
\description{
These data were created to mimic graduate application data.
}
\keyword{datasets}
