% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode variables based on a pre-defined mapping table.}
\usage{
recode(x, table, verbose = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{table}{a data.frame with three columns: variable, category, new_category}
}
\value{
a recoded data.frame.
}
\description{
Recode variables based on a pre-defined mapping table.
}
\examples{

# the main data frame to be recoded
x <- data.frame(
  id = 1:10,
  gender = sample(c("m", "f", "fe"), 10, replace = TRUE),
  labour = sample(c("employed, fulltime", "employed, part-time", "unemployed"), 10, replace = TRUE)
)

# create a mapping table
my_table <- data.frame(variable = c(rep("gender",3), rep("labour", 3)),
                       category = c("m", "f", "fe",
                                    "employed, fulltime", "employed, part-time", "unemployed"),
                       new_category = c("male", "female", "female",
                                        "employed", "employed", "unemployed"))

x_new <- recodeR::recode(x = x, table = my_table, verbose = TRUE)
}
