% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeR-package.R
\docType{package}
\name{recodeR-package}
\alias{recodeR}
\alias{recodeR-package}
\title{recodeR: Package For Recoding Variable Values}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
