% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dict.R
\name{check_dict}
\alias{check_dict}
\title{Check Data Dictionary Structure}
\usage{
check_dict(
  df,
  var_col = "var",
  values_col = "values",
  labels_col = "labels",
  default_col = "default",
  string_col = "string",
  remove_na = FALSE,
  remove_empty = FALSE
)
}
\arguments{
\item{df}{A data.frame or tibble containing unique variable values and value labels
for variables in a dataset. Two optional columns can also be included: `default`, a
default value for old values without a new mapped value/label and `string`, a logical
column indicating whether the new values to be mapped onto the variable are string (1)
or not (0).}

\item{var_col}{Character string of the name of the column in `df` that contains unique
variable/column names in your dataset.}

\item{values_col}{A character string specifying the name of the column in `df` that holds
the unique values for each variable/column in your dataset.}

\item{labels_col}{A character string specifying the name of the column in `df` that holds
the unique values/labels to replace old values for each variable/column in your dataset.}

\item{default_col}{A character string specifying the name of the column in `df` that holds
default values to use when a matching (old) value is not found.}

\item{string_col}{A character string specifying the name of the column in `df` that
indicates whether the new values to be mapped onto the variable are string.}

\item{remove_na}{A logical value indicating whether to remove NA values from values/labels.
Default is `FALSE`.}

\item{remove_empty}{A logical value indicating whether to remove empty values from values/
labels. Default is `FALSE`.}
}
\value{
A tibble
}
\description{
Checks whether (old) values and (new) values/labels for each unique
columns/variable in a data dictionary (flat file) are the same length and unique.
}
