% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recode_dict.R
\name{create_recode_dict}
\alias{create_recode_dict}
\title{Create a 'recoding' data dictionary}
\usage{
create_recode_dict(
  df,
  dict_type = "all",
  var_col = "var",
  from_col = "values",
  to_col = "labels",
  default_col = "default",
  string_col = "string"
)
}
\arguments{
\item{df}{A data frame that lists the unique values and corresponding labels for variables 
in a dataset. It should also include two additional columns: `default`, which specifies a 
fallback value for any unmatched original values, and `string`, a logical indicator denoting 
whether the new mapped values are strings (1) or not (0).}

\item{dict_type}{A character string of the type of dictionary to create. Must be one of 
`all`, `string`, or `numeric`. Default is `all`.}

\item{var_col}{A character string of the name of the column in `df` that contains unique
variable names in your dataset.}

\item{from_col}{A character string specifying the name of the column in `df` that contains 
the original values (`from_values`) for each variable in your dataset.}

\item{to_col}{A character string specifying the name of the column in `df` that contains 
the new values (`to_values`) for each variable in your dataset.}

\item{default_col}{A character string specifying the name of the column in `df` that contains
a fallback value for any unmatched original values.}

\item{string_col}{A character string specifying the name of the column in `df` that indicates 
whether the new values to be mapped onto the variable are string.}
}
\value{
A tibble
}
\description{
Generates a 'recoding' dictionary for both string and numeric variables. 
Each dictionary contains unique variable names from a dataset that require recoding, 
a set of two-sided formulas for mapping old values to new ones, default values for 
unmatched cases, and a flag indicating if the new values for each column are strings 
(1) or not (0).
}
