% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_delim_dict.R
\name{expand_delim_dict}
\alias{expand_delim_dict}
\title{Expand a Data Dictionary}
\usage{
expand_delim_dict(
  df,
  var_col = "var",
  values_col = "values",
  labels_col = "labels",
  default_col = "default",
  string_col = "string",
  value_delim = ",",
  label_delim = ";",
  remove_na = FALSE,
  remove_empty = FALSE,
  check_df = TRUE
)
}
\arguments{
\item{df}{A data.frame or tibble containing unique variable values and value labels
for variables in a dataset. Two optional columns can also be included: `default`, a
default value for old values without a new mapped value/label and `string`, a logical
column indicating whether the new values to be mapped onto the variable are string (1)
or not (0).}

\item{var_col}{Character string of the name of the column in `df` that contains unique
variable/column names in your dataset.}

\item{values_col}{A character string specifying the name of the column in `df` that holds
the unique values for each variable/column in your dataset.}

\item{labels_col}{A character string specifying the name of the column in `df` that holds
the unique values/labels to replace old values for each variable/column in your dataset.}

\item{default_col}{A character string specifying the name of the column in `df` that holds
default values to use when a matching (old) value is not found.}

\item{string_col}{A character string specifying the name of the column in `df` that
indicates whether the new values to be mapped onto the variable are string.}

\item{value_delim}{A character or set of characters used to separate values in the
`values_col`. Default is a comma `,`.}

\item{remove_na}{A logical value indicating whether to remove NA values from values/labels.
Default is `FALSE`.}

\item{remove_empty}{A logical value indicating whether to remove empty values from values/
labels. Default is `FALSE`.}

\item{check_df}{A logical value indicating whether to use the `check_dict` function to verify
that the expanded data dictionary meets specific criteria necessary for use with the `create_dict`
function. Default is `TRUE`.}

\item{labels_delim}{A character or set of characters used to separate values in the
`labels_col`. Default is a semi-colon `;`.}
}
\value{
A tibble
}
\description{
Expands a delimited data dictionary where each row contains unique
variable/column names, (original) values/labels, (new) values/labels, default
values to apply when , and an indicator column identifying whether the new values
to be mapped onto the variable are string. The function transforms the data into
a structured format where each unique value/label pair for each unique column is
associated with a row in the data dictionary.
}
