% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_cols.R
\name{recode_cols}
\alias{recode_cols}
\title{Recode variable values}
\usage{
recode_cols(
  df,
  list_of_dicts,
  var_col = "var",
  formula_pairs_col = "formula_pairs",
  default_col = "default",
  string_col = "string"
)
}
\arguments{
\item{df}{A data frame.}

\item{list_of_dicts}{A list of 'recoding' dictionaries.}

\item{var_col}{A character string of the name of the column in `list_of_dicts` that contains 
unique variable names in your dataset.}

\item{formula_pairs_col}{A character string of the name of the column in `list_of_dicts` that
contains a sequence of two-sided formulas for recoding variable values.}

\item{default_col}{A character string of the name of the column in `list_of_dicts` that contains
a fallback value for any unmatched original values.}

\item{string_col}{A character string of the name of the column in `list_of_dicts` that
indicates whether the new values to be mapped onto the variable are string.}
}
\value{
A tibble
}
\description{
`recode_cols()` recodes the values of variables in a data frame using a 'recoding' 
dictionary, which comprises of a series of two-sided formulas to define a mapping from old values 
to new ones
}
