% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeR-package.R
\docType{package}
\name{recodeR-package}
\alias{recodeR}
\alias{recodeR-package}
\title{recodeR: Recode Variables in A Data Frame With A Pre-Defined Mapping Table}
\description{
The goal of recodeR is to make recoding super easy when you have a pre-defined mapping table. 
    I often found myself using the dplyr::case_when function in data prepation and it can look super 
    messy real fast when you have a lot of categories to be recoded. Hence, recodeR was born!
}
