\name{recoder-package}
\alias{recoder-package}
\docType{package}
\title{A Simple and Flexible Recoder}

\description{
Simple, easy to use, and quite flexible function for recoding variables. Among the features is that it can handle compound conditions and can include functional transformations on the go, which together allow for rapid piecewise definition of functions.}
\details{
\tabular{ll}{
Package: \tab recoder\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-08\cr
License: \tab GPL (>= 2) \cr
}
}

\author{
Ali Sanaei}

\keyword{package}
